refactor: auctions에서 검색 기록 로직 분리 후 search 패키지로 수정
