+++ src/main/java/com/samyookgoo/palgoosam/search/dto/SearchHistoryCreateRequestDto.java
@@ -1,4 +1,4 @@
-package com.samyookgoo.palgoosam.auction.dto;
+package com.samyookgoo.palgoosam.search.dto;
 
 import lombok.Getter;
 import lombok.Setter;
