+++ src/main/java/com/samyookgoo/palgoosam/search/repository/SearchHistoryRepository.java
@@ -1,6 +1,6 @@
-package com.samyookgoo.palgoosam.auction.repository;
+package com.samyookgoo.palgoosam.search.repository;
 
-import com.samyookgoo.palgoosam.auction.domain.SearchHistory;
+import com.samyookgoo.palgoosam.search.domain.SearchHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
