+++ src/main/java/com/samyookgoo/palgoosam/search/service/SearchHistoryService.java
@@ -1,0 +1,53 @@
+package com.samyookgoo.palgoosam.search.service;
+
+import com.samyookgoo.palgoosam.search.domain.SearchHistory;
+import com.samyookgoo.palgoosam.search.dto.SearchHistoryCreateRequestDto;
+import com.samyookgoo.palgoosam.search.repository.SearchHistoryRepository;
+import com.samyookgoo.palgoosam.user.domain.User;
+import com.samyookgoo.palgoosam.user.repository.UserRepository;
+import jakarta.persistence.EntityNotFoundException;
+import java.util.Optional;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+@Service
+@Slf4j
+@RequiredArgsConstructor
+public class SearchHistoryService {
+
+    private final SearchHistoryRepository searchHistoryRepository;
+    private final UserRepository userRepository;
+
+    public void recordUserSearch(SearchHistoryCreateRequestDto requestDto) {
+        /*
+        userService에서 사용자 판별하는 부분 추가
+        if (user != null) 아래 로직 동작하도록
+         */
+        User dummyUser = userRepository.findById(1L).orElseThrow(() -> new EntityNotFoundException("User not found"));
+        Optional<SearchHistory> processed = processSearchKeyword(requestDto, dummyUser);
+        processed.ifPresent(searchHistoryRepository::save);
+    }
+
+
+    private Optional<SearchHistory> processSearchKeyword(SearchHistoryCreateRequestDto requestDto, User user) {
+        String keyword = requestDto.getKeyword().trim();
+        if (keyword.isEmpty()) {
+            return Optional.empty();
+        }
+        SearchHistory existingSearch = searchHistoryRepository.findByKeywordAndUserAndIsDeleted(
+                keyword,
+                user.getId(), false);
+
+        if (existingSearch != null) {
+            existingSearch.incrementSearchCount();
+            return Optional.of(existingSearch);
+        }
+
+        return Optional.of(SearchHistory.builder()
+                .keyword(keyword)
+                .user(user)
+                .searchCount(1L)
+                .build());
+    }
+}
